move pkg/cryptoinfo
